remote: refactor the code, more classic help and controls
Stop with Backspace, quit with ^D (standard for programs which read
input from the terminal).
Invert song and volume control, it seems more natural.
AZERTY and QWERTY.

Add some options.
* Standard help message.
* Can specify a pipe to write to.

Perform some validity tests on command-line arguments and environment,
exit cleanly.
No more error output when a wrong key is pressed in the terminal, it
was an annoyance.

Remove the hack with $i < 1 to break the loop.

Be careful with the UTF-8 BOM (first two bytes of the file, U+FEFF) !
It messed up the shebang, bash was complaining about that first line.